refactor: moved root command creation in types/cmd
